Monet-2/requirements.txt
sh==2.2.2
requests==2.32.3
GitPython==3.1.44
markdown==3.7
Werkzeug==3.1.3
numpy==2.2.4
flask==3.1.0
Flask-Caching==2.3.0
authlib==1.3.2
filelock==3.16.1
bokeh==3.7.2
typing-extensions==4.13.0
pytz==2025.2
python-dateutil==2.9.0.post0
gunicorn==23.0.0
tinydb==4.8.2
apscheduler==3.11.0
pyjwt==2.10.1
mysql-connector-python==9.2.0
flask-cors==5.0.1
httpcore==1.0.7
colorhash==2.0.0
flask-assets==2.1.0
cssmin==0.2.0
rjsmin==1.2.4
sqlalchemy
pydantic